bug fix in load balancer for sending data

+++ rpc_structs/rpc_structs.go
@@ -2,8 +2,8 @@ package rpcstructs
 
 type Args struct {
 	JobId               int
-	CPUResourceUsage    int
-	MemoryResourceUsage int
+	CPUResourceUsage    float64
+	MemoryResourceUsage float64
 	TimeStart           int
 	TimeEnd             int
 	TaskId              int
